Catch exceptions on get hinge (#2662)
https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3094140

In certain cases, the getHinge method triggers a NullPointerException,
resulting in a crash on Edge. To handle this more gracefully, instead of
throwing the exception, we are returning the default Rect.

The full solution involves replacing DisplayMask with Jetpack Window
Manager. However, for this release, we are implementing a temporary
patch, with the complete fix planned for the next release.



https://learn.microsoft.com/en-us/previous-versions/dual-screen/android/api-reference/dualscreen-library/core/screen-info