Fixing the sign in screens when edge to edge is enabled, Fixes AB#3293334 (#2665)
The Auth password page on LTW is edge-to-edge by default on devices
running Android 15 if the app is targeting Android 15 (API level 35).
Please refer to
[this](https://developer.android.com/about/versions/15/behavior-changes-15#edge-to-edge)
for more details. Because of this feature enabled by default, I am
seeing below cropped screen.

Attached screenshot below.

![EdgeToEdge](https://github.com/user-attachments/assets/a3423b03-336d-45f8-ba27-647faa8ae65e)

Steps to reproduce:
1. set targetSdkVersion to 35 in BrokerHost app

2. launch MsalTestApp

3. click on Sign in button

4. See that the page is cropped on the top
-------------------------------------------------------------
Expected Behavior:
Should not be cropped

Fix : Adding padding in the activity showing the UI screens by getting
the insets of the system bar as suggested in the edge to edge
documentation.

This is after fixing the issue

![image](https://github.com/user-attachments/assets/6637bd39-cdda-4701-a0bf-7ea1740c4d38)

Fixes
[AB#3293334](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3293334)