Showing webcp flow in webview , Fixes AB#3135912 (#2673)
Intune's support for web enrollment for Android is scheduled to GA in Q3
CY2025.
For the many customers who set up conditional access, users enter the
web enrollment flow by being blocked from accessing a productivity app
such as Teams or Outlook.

From that block page, they are pushed out of their productivity app and
sent to the default browser app on their Android device where they see a
Company Portal website (webCP) page, which guides them to Settings where
they can enroll in Intune management.

**Problems**

The problems with showing webCP in the browser app is:

1. Additional sign in: During enrollment, it adds an additional
authentication for the user (2nd auth for iOS and 3rd for Android).
2. Context change: It is a jarring experience, taking the user out of
the context of the productivity app and into the browser app, which can
be disorienting.

See snapshot below for current experience with WebCP flow enabled

![image](https://github.com/user-attachments/assets/7ea1ce0b-10c9-4592-b659-8abd9e40ce59)

Note, the highlighted green pages indicate the 3 times the user must
authenticate during this flow.

**The flow:**

1. The user signs-in to the productivity app and is blocked by MDM CA.
2. The user must select the “Continue” button on the MDM CA page (which
indicates they need to register their device).
3. The user is navigated to WebCP in the browser, where they must
sign-in again.
4. The user is shown a page to kick-off WP enrollment directly from the
browser. The user must select the button to “Get started”.
5. The user is taken through the Google screens for creating the Work
Profile.
6. The user lands
7. The user completes the enrollment flow in the Work Profile version of
the Company Portal.

**Fix**
Showing webcp in webview when user clicks on continue button on CA block
page.
So with the fix, the screen shown in red in step 2 in below snapshot is
removed

![image](https://github.com/user-attachments/assets/0f012cb8-d495-40e0-a235-f3032bd93b78)

Fixes
[AB#3135912](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3135912)

Related broker PR :
https://github.com/AzureAD/ad-accounts-for-android/pull/3130

---------

Co-authored-by: Shahzaib <37125644+shahzaibj@users.noreply.github.com>