Fix playstore url launch for broker apps, Fixes AB#3307594 (#2689)
Issue : When there is an intent to launch to playstore URL for broker
apps, webview loads the page
https://play.google.com/store/apps/details?id=com.microsoft.windowsintune.companyportal
for a brief second and then redirects to the intent:// which gets
launched using
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/blob/e5b119cb218560210befbf570b89b3f4271def9c/common/src/main/java/com/microsoft/identity/common/internal/ui/webview/AzureActiveDirectoryWebViewClient.java#L778
In this PR, I am fixing the momentary load of the URL and directly
launching the intent for a smooth experience.

- I put the change behind a flight which is ON by default.
- If for some reason this fails by throwing an exception, I am returning
false so that webview loads the URL and follows the existing path
instead of completely breaking.

Fixes
[AB#3307594](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3307594)