Flighting webcp in webview, Fixes AB#3307594 (#2686)
1. Flighting for webcp in webview is divided into 3 parts
- Not flighting this feature for brokerless scenarios. The flight will
be off by default for brokerless scenarios as Intune team is ok to only
target brokered flows since this is the majority case.
- Now coming to brokered scenarios, we have below
- Simple flighting with ENABLE_WEB_CP_IN_WEBVIEW which will be used for
rollout starting October.
- Using tenantId list TENANT_LIST_TO_ENABLE_WEB_CP_IN_WEBVIEW. This
flight will be used first in the months of August and September. This
will be 100% rolled out in the beginning itself. This flight will hold
upto 25 tenantIds. See below snapshot

![image](https://github.com/user-attachments/assets/4eff53a6-3977-4ba2-88c0-5529fe5f69c9)


2. I intentionally added a global variable to keep the flight value
instead of calling isWebCpInWebviewFeatureEnabled method because I do
not want to invoke code to get tenantId multiple times. Also, this value
must remain the same once the webcp url is loaded in webview.

3. Apart from this, I have added a special method to handle google
enrollment URL which has extra intent flags needed for this scenario.

Related PR :
https://github.com/AzureAD/ad-accounts-for-android/pull/3140

Fixes
[AB#3307594](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3307594)