Edge to edge fix limited to brokered only scenarios, Fixes AB#3293334 (#2690)
The fix I did in this PR is flighted
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2665
But we cannot control flighting when no broker is installed. I have
checked with OneAuth team and they are ok with just turning this fix ON
for brokered flow since that is the majority case.

Fix : Simply limit the fix to brokered flow by setting the default
flight value in common to false and in broker to true. See broker PR :
https://github.com/AzureAD/ad-accounts-for-android/pull/3130

Fixes
[AB#3293334](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3293334)