Updating handling of ssl error received in Android WebView's onReceivedSslError callback, Fixes AB#3268908 (#2691)
Update:
After talking to @iulico-1 from OneAuth.
1. First rolling out with Broker and keeping disabled for OneAuth (and
MSAL Android). Added SSL error telemetry
2. Enable for oneauth after we see no issues in Broker. So, the toggle
code to let OneAuth enable/disable is removed from the PR.
3. Investigate how we can pass the error to OneAuth to track SSL failure
and User canceled the flow because of SSL error.

=====================

Fixes
[AB#3268908](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3268908)

**Summary:**  
This PR updates the handling of SSL errors in the Android WebView used
for authentication. The change is in response to an incident where the
MSA sign-up flow failed due to an expired certificate on an MSA UX web
page. The expired certificate was associated with a resource loaded by
the page; however, the existing SSL error handling in Common’s WebView
would cancel the loading of that resource and subsequently terminate the
entire authentication flow. (See PBI for details.)

**Solution**
**Option 1**
We simply log and stop loading the resource that caused the error by
calling `SssHandler.cancel()` but not stop the flow by erroring out.
This is a simple change. The only thing is the main URL (the url loading
in the WebView) itself has SSL error then the flow would end up in
*blank page*. This is as bad as finishing the flow but does not inform
calling app of the error via exception.

But if sub-resource has the error, then the flow continues.

**Flighting**
**New SSL Error Handler**: 
- ~~Adds a new intent key (`WEB_VIEW_NEW_SSL_ERROR_HANDLER_ENABLED`) in
`AuthenticationConstants.java` to control whether the new SSL error
handler logic is enabled. This can be used by OneAuth to opt-in or
opt-out. If the key is not passed, then Common's default logic based on
flighting will be used. If the key passed, that will be used (primarily
meant for OneAuth/MSALCPP).~~. The toggle would be added later if
required. This simplifies current logic to only rely on Common/broker's
flighting.
- Adds a new flight value is well defaulting to `false`. So, OneAuth and
MSAL would also use default flight value and continue using current
behavior. For Broker, we would enable new flow with default `true`
(https://github.com/AzureAD/ad-accounts-for-android/pull/3141), which
can be controlled remotely to turn off if required.

**Option 2** (Not this PR, added for reference)
The objective of this update is to improve SSL error handling so that
the flow is only cancelled if an SSL error occurs on the main frame
resource (the primary page being loaded). If an SSL error affects a
sub-resource, only the loading of that specific resource is cancelled,
and the overall authentication flow continues. This approach helps
prevent unnecessary failures caused by SSL errors on secondary or
non-critical resources.
(https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2688)

The con of the approach is client needs to track the active url that's
about and being loaded in WebView. The tracking adds cost of developing
and maintaining to solution without significant benefit over option 1.