Using Baggage to propagate attributes from parent Span, Fixes AB#3262849 (#2671)
### Summary
The passkey telemetry dashboard (and maybe some others) has had issues
with latency due to joins being done between the Fido span and the
parent span related to acquire token interactive. These join statements
by trace id allowed us to associate attributes (such as tenant id) to
the Fido span that weren't available in the context of the Fido classes.
In order to directly save the attributes onto the Fido span itself, we
can make use of [Otel's
Baggage](https://opentelemetry.io/docs/specs/otel/baggage/api/), which
is meant to help hold data across multiple spans (from parent to
children). We can then use an implementation of [Otel's
TextMapPropagator](https://opentelemetry.io/docs/specs/otel/context/api-propagators/)
to serialize the Otel Context and pass it to the AuthenticationActivity
as an Extra (similar to what we are doing for spanContext). This Context
can then be accessed by the WebViewClient and Fido related classes,
where they will retrieve the Baggage.

This PR has the changes to implement this behavior and adds a few static
helper classes for TextMapPropagator and Baggage (the baggage class is
more relevant for the corresponding Broker PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/3119).

While this PR tackles attribute propagation for AuthorizationActivity
specifically, Baggage and the helper classes can be used in other areas
of the code where propagation is needed as well.


[AB#3262849](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3262849)