Fix for enrollment id not sent to /token from local controller (#2726)
Fix for ICM:
https://portal.microsofticm.com/imp/v5/incidents/details/663024585/summary

What happens is follows:

- MAM compliant app makes ATS call to Broker
- Broker doesn't have PRT for account so fallbacks to local RT
(BrokerLocalController)
- Broker correctly queries MAM enrollment id from CP content providers
and populates it on CommandParameters
- Now request goes through BrokerLocalController which apparently uses
BaseController (as opposed to AbstractBrokerController) for creating
token request to renew AT with RT.
- BaseController ignores the MAM enrollment id from CommandParameters
and doesn't attach it to the request.
- ESTS throws IntuneAppProtectionPolicyRequired error because it doesn't
see enrollment ID in request

Had BrokerLocalController actually relied on its immediate parent class
i.e. AbstractBrokerLocalController this would've worked but that's not
the case.

Changing the above is a big refactor.

To fix the issue, I'm just attaching MAM enrollment ID (if present on
CommandParameters) to the request in BaseController.