Securing handling added for webapps with webcp flight, Fixes AB#3344894 (#2733)
In this PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2732,
I added a new check to detect device CA flow and follow our current
pattern of handling the device CA requests. But I did not secure it with
a flight.

There are 2 ways of adding a flight for this
1. Add a new flight that is independent of webcp feature 
2. Depend on webcp feature only. 
I went with 2nd option, because I only really need these changes when
webcp flight is ON so that users are not blocked by new flow. So, I am
securing it with webcp flight.


Fixes
[AB#3344894](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3344894)