Handle BadTokenException gracefully in CBA dialogs, Fixes AB#3346402 (#2731)
### Summary
Fixes
[AB#3346402](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3346402)

If activities are destroyed or in the process of being destroyed while a
CBA dialog is going to pop up, a BadTokenException will get thrown, and
this is currently not being handled by the CBA logic. This PR adds
try/catch clauses to the "show" methods of the DialogHolder, which log
the exception and call the appropriate callback methods to finish out
the CBA flow as a cancellation.