[Common] Update Broker ATS for Resource Accounts, Fixes AB#3230426 (#2704)
Don't return UiRequiredException when the acquire token silent is for
resource account. It's not expected to start interactive request for an
RA account, so instead we return ServiceException.

On top of it, there will be broker side PR to handle interrupt
internally if ATS fails with relevant errors. This will be separate
change.

Fixes
[AB#3230426](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3230426)