Fix for App Link Usage in DUNA / SSO scenarios (#2745)
**Issue:** When app link redirect URI is used from within CCT then CCT
may not redirect to the native app to handle the redirect and instead
open it directly within the browser. From experience I've seen this
happen there are existing cookies present at the IdP and there is no
user interaction.

**Solution:** Per [this chromium
thread](https://issues.chromium.org/issues/40918490), the solution is
set a flag when launching CCT called as
[EXTRA_SEND_TO_EXTERNAL_DEFAULT_HANDLER](https://developer.android.com/reference/kotlin/androidx/browser/customtabs/CustomTabsIntent#EXTRA_SEND_TO_EXTERNAL_DEFAULT_HANDLER()).

That flag was actually introduced in `androidx.browser` library version
`1.7.0` and we were on a very ancient version (1.0.0). So this PR
updates that library version to `1.7.0`

As a side effect of that upgrade we can no longer use
LocalBroadcastManager API (it's deprecated). So I've updated that
accordingly similar to what @p3dr0rv was doing earlier when he tried
upgrading browser library to 1.9.0 (which we couldn't do due to
requiring compile sdk 36)