Added error handling when webcp redirects have browser protocol, Fixes AB#3385536 (#2767)
Issue : There is a recent change made in WebCP side to pass browser
protocol redirects for China market urls, 3rd part MDM urls and also the
Url to playstore launch for installing CP. This is an unexpected change
on their side and we did not have proper handling for these cases.

Fix : 
1. Added try/catch block around the operations in processWebsiteRequest
method.
2. Returning MDM_FLOW as the result instead of CANCELLED result
currently being returned when we open link in a browser for webcp urls.
3. Added telemetry for operations in processWebsiteRequest method.
4. Made some minor improvements in AzureActiveDirectoryWebViewClient
class to avoid duplicate code for getting broker package name from url.

Related broker with just an attribute added :
https://github.com/AzureAD/ad-accounts-for-android/pull/3228

Fixes
[AB#3385536](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3385536)