Use baggage for spans, Fixes AB#3385536 (#2770)
Using baggage to get parent span's (AcquireTokenInteractive)
correlationId, calling app package name and setting them on the child
spans created during webcp redirects in webview.

Why this needed?
- We do not have correlationId set on webcp spans as of today. Although,
we can check for the parent span and get correlationId from there, it is
just an additional step. I want to avoid this.
- We cannot directly set attributes to indicate that is webcp flow on
the parent span (ATI) since the context passed to the activity is
immutable. Recommended way is to create a child span and set attributes
as needed on it.
- We also cannot get parent span's attributes on child span directly
using the context passed. Support for this was added by using baggage in
this PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2671.
So, I am leveraging this.


Related broker PR to remove the redundant attributes :
https://github.com/AzureAD/ad-accounts-for-android/pull/3230

Fixes
[AB#3385536](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3385536)