Fix for app link redirect from CCT due to forced browser preference, Fixes AB#3387976 (#2775)
Fixes
[AB#3387976](https://dev.azure.com/IdentityDivision/Engineering/_workitems/edit/3387976)

The previous fix seems to have addressed the issue only partially. 
Previous fix:
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2745

Basically what's happening is that even with that flag CCT will
sometimes still launch the URL in browser and this is happening because
we're setting the browser's pkg name explicitly on the intent which
tells the OS that this MUST be opened in the browser.

Checked CCT logs and they print something like:

`Launching in browser as Browser pkg was explicitly set on the Intent`

Once we remove the explicit pkg name then OS finds the best possible app
on the device to handle the intent. For an app link that's the the
broker hosting app.