Update Java Script Api Exposure, Fixes AB#3385083 (#2773)
This PR will move the code that exposes the java script api from
handleUrl on pageStarted. This will make it so we use the url that we
are actually planning to load in the webview to validate whether or not
to exposue the JS interface. previously, most pages were being validated
correctly in handleUrl, but in certain redirects, we may have been
missing some urls that should have had the url exposed. This change
addresses those gaps by checking every url the webview is about to load.


[AB#3385083](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3385083)