getAllSsoTokens method for Edge, Fixes AB#3361797 (#2774)
### Summary
We have a getSsoToken for OneAuth/Edge which returns a PRT cookie
corresponding to a given account. They have asked us to also implement a
method which gets all (well, we're capping it at the 3 most recent) PRT
cookies based on Environment/cloud, which we can get from a passed in
Authority value.
The changes in common for this method involve defining the method in
BrokerMsalController, updating the appropriate adapter classes, creating
the AcquirePrtSsoTokenBatchResult class, and bumping the broker protocol
version.

Corresponding Broker PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/3233

Fixes
[AB#3361797](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3361797)