Web Apps account id registry, Fixes AB#3406762 (#2787)
[AB#3406762](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3406762)
### Summary
The WebAppsAccountIdRegistry maps account ids to sets of client ids.
This class is created for the purpose of the Edge TB APIs work, where it
will be used for the getToken operation (silent case where we want to
check that the user has signed into this webapp before) and the SignOut
operation (to remember which web apps we need to sign out of).

Broker PR for the SignOut operation using this registry will follow.