partial revert to fix failing tests in broker FixesAB#3397557, Fixes AB#3397557 (#2797)
Fixes[AB#3397557](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3397557)

Recently refactored getAccountByLocalAccountId method to first filter
both accounts and credentials on environment and then, check their
homeAccountId to match the accounts with credentials. However, that
refactoring does not work when environment passed is **null**. This
resulted in test failures in the broker repo.
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2781
 
This PR is to revert some of the changes made as part of that commit and
fix the failing tests.