Edge TB: Expose WebApps APIs, Fixes AB#3409879 (#2793)
### Summary:

In this PR, we:
- Expose the WebApp APIs via BrokerMsalController. OneAuth is already
using this class to call similar Edge APIs.
- getSupportedWebAppContracts(String minBrokerProtocolVersion): this
method is called by OneAuth to make sure that the broker can handle the
request.
- executeWebAppRequest(String Request, String minBrokerProtocolVersion):
this method handles requests for GetToken, GetCookies, and SignOut.
- Implement the bundle helper methods for creating the bundle and
extracting the results from the result bundle.
- Adds the bundle keys to AuthenticationConstants, where both common and
broker can access them.

Related broker PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/3256


[AB#3409879](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3409879/)