Fix for SDL violation in device pop scenarios, Fixes AB#3284510 (#2744)
- 1 of the SDL assessments has a requirement that we should avoid using
multiple crypto purposes with the same key.
- There are multiple places where we are using multiple purposes and
device pop scenario is one of them.
- In Device pop scenario, we do not use encrypt/decrypt methods anywhere
in our code. Only signing is used. Hence removing the ununsed purposes
encrypt, decrypt from the key gen spec.
- Also, encrypt and decrypt methods added in
[IDevicePopManager.java](https://github.com/AzureAD/microsoft-authentication-library-common-for-android/compare/somalaya/devicePopSdlFix?expand=1#diff-2ca2fe49df086f6ee051aeeedcc5bf3bfd53b1b3390e2a8d1143af2cfc74da7e)
are probably for maintaining all the crypto operations but I have
deprecated them as they are not getting used.
- Added telemetry in mintSignedHttpRequest method. 

Note : The flight to remove encryption paddings in initialize method of
AndroidDevicePopManager is not going to be enabled until we observe
telemetry added in this PR first. Idea is that the telemetry should
prove that encrypt/decrypt methods are not getting used. Only then, we
can enable the flight! So, this is going to be enabled in Jan/Feb next
year.


Fixes
[AB#3284510](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3284510)