Edge TB: changes related to GetToken, Fixes AB#3317102 (#2803)
### Summary

Broker PR: https://github.com/AzureAD/ad-accounts-for-android/pull/3267

Changes related to WebApps getToken are done in this PR, and it also
notably contains refactoring (such as moving some files from broker to
common) and some additions to core classes such as AccountChooser and
AccountChooserActivity.

To optimize the number of IPCs and repeated code, the getToken operation
will be completed as follows:

1. Caller calls BrokerMsalController.ExecuteWebAppsRequest.
2. BrokerMsalController takes the request and calls the broker, which
will start the ExecuteWebAppsRequestOperation.
3. ExecuteWebAppsRequestOperation takes a peek at the method and
delegates to the correct sub operation.
4. For the WebAppsGetTokenSubOperation, determine if silent is possible.
Otherwise, force an interactive request and send an intent back to
BrokerMsalController.
5. If silent is possible, try a silent request. If a token is returned,
this will get formed into a response and sent back to
BrokerMsalController. If an exception is thrown, the operation
determines if an interactive request should be made. If true, an intent
is sent back. If false, the exception is formed into a response and sent
back to BrokerMsalController.
6. Back at BrokerMsalController, if a token response or error response
is received, this response string is returned as-is. If an intent is
received, BrokerMsalController will start the intent (you'll see some
refactoring with the existing acquireToken method to help reusing
logic).
7. AccountChooserActivity will handle the request and knows that it is
from a webapp. It will return either a token response or an error
response in the protocol format.
8. BrokerMsalController returns this response as-is.


[AB#3317102](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3317102)